Cleanup and some refactoring
